requirements.txt
# python ~= 3.11
pip==23.0
numpy~=1.24.2
pandas==1.5.3
openpyxl==3.1.0
et-xmlfile==1.1.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
six==1.16.0